requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
Flask==3.1.0
flatbuffers==25.2.10
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
onnxruntime==1.21.1
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
protobuf==6.30.2
PyMatting==1.1.13
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.2
sympy==1.14.0
tifffile==2025.3.30
tqdm==4.67.1
urllib3==2.4.0
Werkzeug==3.1.3
